refactor(analysis): extract resolveImportPath to BaseLanguageAnalyzer
