+++ analysis/src/main/kotlin/de/maibornwolff/dependacharta/pipeline/analysis/analyzers/typescript/TypescriptAnalyzer.kt
@@ -1,7 +1,6 @@
 package de.maibornwolff.dependacharta.pipeline.analysis.analyzers.typescript
 
 import de.maibornwolff.dependacharta.pipeline.analysis.analyzers.BaseLanguageAnalyzer
-import de.maibornwolff.dependacharta.pipeline.analysis.analyzers.common.utils.stripSourceFileExtension
 import de.maibornwolff.dependacharta.pipeline.analysis.analyzers.common.utils.withoutFileSuffix
 import de.maibornwolff.dependacharta.pipeline.analysis.model.Dependency
 import de.maibornwolff.dependacharta.pipeline.analysis.model.FileInfo
@@ -14,7 +13,7 @@ import de.maibornwolff.treesitter.excavationsite.api.Language
 const val DEFAULT_EXPORT_NODE_NAME = "DEFAULT_EXPORT"
 
 class TypescriptAnalyzer(
-    private val fileInfo: FileInfo,
+    fileInfo: FileInfo,
 ) : BaseLanguageAnalyzer(fileInfo) {
     override val language = SupportedLanguage.TYPESCRIPT
 
@@ -31,17 +30,5 @@ class TypescriptAnalyzer(
     override fun convertImport(import: ImportDeclaration): Dependency {
         val resolvedPath = resolveImportPath(import.path)
         return Dependency(path = Path(resolvedPath), isWildcard = import.isWildcard)
-    }
-
-    private fun resolveImportPath(tsePath: List<String>): List<String> {
-        val stripped = tsePath.map { it.stripSourceFileExtension() }.filter { it.isNotEmpty() }
-        if (stripped.isEmpty() || (stripped.first() != "." && stripped.first() != "..")) return stripped
-        var dirParts = fileInfo.physicalPathAsPath().parts.dropLast(1)
-        var remaining = stripped
-        while (remaining.isNotEmpty() && (remaining.first() == "." || remaining.first() == "..")) {
-            if (remaining.first() == "..") dirParts = dirParts.dropLast(1)
-            remaining = remaining.drop(1)
-        }
-        return dirParts + remaining
     }
 }
